refactor(auth): remove old authentication components and implement new layout and forms
